+++ services/bls/model_disable_alarm_policy_request.go
@@ -1,5 +1,5 @@
 package bls
 
 type DisableAlarmPolicyRequest struct {
-	Name []*string `json:"name,omitempty"`
+	Names []*string `json:"names,omitempty"`
 }