+++ services/bls/model_enable_alarm_policy_request.go
@@ -1,5 +1,5 @@
 package bls
 
 type EnableAlarmPolicyRequest struct {
-	Name []*string `json:"name,omitempty"`
+	Names []*string `json:"names,omitempty"`
 }